[refac] move JwtAuthenticationEntryPoint.java
